DATAMONGO-1165 - Add support for Streaming large result lists.
WIP
