+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/repository/support/SpringDataMongodbSerializer.java
@@ -100,7 +100,7 @@ class SpringDataMongodbSerializer extends MongodbSerializer {
 	protected DBObject asDBObject(String key, Object value) {
 
 		if (ID_KEY.equals(key)) {
-			return mapper.getMappedObject(super.asDBObject(key, value), null);
+			return mapper.getMappedObject(super.asDBObject(key, value), (MongoPersistentEntity<?>) null);
 		}
 
 		return super.asDBObject(key, value instanceof Pattern ? value : converter.convertToMongoType(value));
